empresa

+++ src/models/Stock.js
@@ -8,7 +8,7 @@ const Stock = sequelize.define(
     {
         id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
         productId: { type: DataTypes.INTEGER, allowNull: false },
-        quantity: { type: DataTypes.INTEGER, allowNull: false },
+        quantity: { type: DataTypes.FLOAT, allowNull: false },
     },
     { timestamps: true }
 );
